Remove animation implementation (#28)
Hyperview had an old animation implementation that wasn't fully supported. This was designed before the full behavior system. With behaviors, there's a better path for supporting animations in the future, so we can ditch the old system.

Also: fixed prettier config to not change double quotes to single quotes in Markdown files. This was breaking document generation.