Extract XML fetching code in its own service Goals of this code change: - centralize code that loads remote content, currently spread out between `load` and `fetchElement` - use async/await style instead of chained promise to ease readability and avoid variable mutations - in a future PR, have a better handling for error responses that can containt non-xml content type (i.e. Django errors responses are text/html) - more tests and flow coverage (src/index.js is explicitely ignored by linters as it contains too many issues)
Regarding the naming of the service, I choose `dom` even if right now it has little to do with DOM manipulation. The reasoning is to have a common place for abstractions of the `xmldom` module, and the goal would be to eventually move most of the DOM helpers currently living in [src/services/index.js (currently used a dumping ground for shared functions)](https://github.com/Instawork/hyperview/blob/master/src/services/index.js#L33-L36). I'm open to suggestions for a better name if any.

Implementation notes:
I choose a instantiable class over a simple function, so that the component props `fetch`, `onParseBefore` and `onParseAfter` could only be injected once.
I've also incorporated in this class the responsability to set the `Accept=applicant/xml` header, hence why it was removed from the demo app.