Hyperview XML Schema (#144)
Adding an XML schema for Hyperview that can be used for validation, docs generation, IDE tools, etc.

- The schema files live under the `schema/` directory. `hyperview.xsd` is the main schema, and it imports other schemas for the various namespaces we use.
- I added a command `yarn test:validate-xml` that validates the schema against the XML files under `examples/`.
- I fixed up the examples that contained easy-to-fix issues like wrong namespaces, missing attributes, etc.

However, some examples cannot be validated due to an issue in the Hyperview spec. The schema revealed a conflict between two features:
- Many elements like `<option>` use the attribute `value` to hold the input value.
- The `set-value` behavior uses the attribute `value` to update any input element.
This means that an `<option>` field that triggers a behavior to set another field is ambiguous, because `value` has two meanings.

We will need to resolve this issue in the spec before completing the schema.