fix: Prevent crashes on invalid markup Note: this is a temporary work around to prevent crashes. Proper fix should be addressed in the form of refactors of faulty components.
This issue occurs as we have numerous places in the code base that expects an element to be returned, even when not present.
e.g. with the following markup:
```xml
<section-list>
  <section>
    <item><text>Foo</text></item>
    <item><text>Bar</text></item>
  </section>
</section-list>
```
in the example above, we're missing the `<section-title>` element, but [this code](https://github.com/Instawork/hyperview/blob/master/src/components/hv-section-list/index.js#L105-L107) assumes it does. Upon trying to render it, the app crashes as we're trying to access properties on an `undefined` object. There's a variety of other places in the code that attempt to access elements without verifying it exist first, so this fix should help act as a "catch-all".
Ideally, proper typing should help prevent these kind of bugs. For the time being, treating the element as an optional when rendering will avoid these crashes.