feature(selection) - Track the current selection state of each nav-route (#623)
Reflect the user's selections in the document

- Use the 'focus' listener to allow each hv-route to set its selected
state within the document
- De-select any sibling routes to ensure the current is the only
selected

Note: I did not use 'blur' to have each route deselect itself as that
would cause routes in the background to lose their state.

In the video, as the user selects between the different tabs, the
various \<nav-route\> children of `tabs-navigator` are shown to reflect
current selection. When the `company` modal is popped up, the `company`
\<nav-route\> shows selected while its sibling `tabs-route` becomes
deselected. Note that the `tabs-navigator` children retain their correct
selection state.

Asana: https://app.asana.com/0/0/1205197138955014/f


https://github.com/Instawork/hyperview/assets/127122858/0687558f-c500-419a-bd69-bbd0843ade72