refactor(hv-date-field): Add modal overlay (#630)
This updates the UX of the date field component, by adding an overlay /
"scrim" element covering the screen under the modal. The overlay is
tappable and dismiss the modal (same action as pressing the "cancel"
button). Both opening and closing action cause the modal to slide up and
down in an animated fashion.

Implementation details:
The slide up animation of the modal is recreated with an animated view -
this is needed so that the overlay can be decoupled from this animation.

Another refactor will follow to pull out the modal + overlay components,
in order to share/reuse these with the picker field.



https://github.com/Instawork/hyperview/assets/309515/f7f0ad93-a31c-417b-857e-b588512e99ca

Co-authored-by: flochtililoch <flochtililoch@gmail.com>