fix: `visible` behavior missing detections (#659)
When toggling views with `show`/`hide` behaviors, if the layout is
identical enough (i.e. same elements structure), then React does not
unmount the instances of `VisibilityDetectingView`, which in turn cause
the previous "visible" state to not be reset. The code change here works
around the issue, by adding an `id` prop to `VisibilityDetectingView`,
which is then used to decide internally to reset the "visible" state.
The `id` prop is either set from an existing `id` attribute on the
element targeted by the `visible` behavior, or generated, using every
behavior attributes available on the element (which means if all
attributes are identical, the observed issue will remain - however
unlikely to happen).

The fix also addresses another bug that was uncovered after the first
resolution: we were relying on an older version of the DOM when
triggering the "visible" callbacks. Instead, query the DOM in the
callback.

Note: the use case "Tabs" in example has also been reworked to help
diagnose the issue.


| Before | After |
|-------|------|
|
![before](https://github.com/Instawork/hyperview/assets/309515/3febfad5-8a19-465d-8efa-805168a68078)
|
![after](https://github.com/Instawork/hyperview/assets/309515/fbffb58b-0d8b-44d4-bd11-94207477ee57)
|

Co-authored-by: flochtililoch <flochtililoch@gmail.com>