refactor: `RefreshControl` context (#673)
`RefreshControl` context allows for customizing the refresh control in
`<list>` and `<section-list>`. Consumers were setup in a React class
component fashion, using the static `contextType` property. This
approach is limiting, as only one context can be set on a given class
component, but is necessary when the context needs to be accessed
outside the render method. Since this is not the case for either of
these classes, refactor to use the JSX consumer style in the render
methods, which now leaves room for introducing another context (future
PR coming soon) that does need to be accessed outside the render method.
Additionally, this re-adds the missing `refreshControl` property being
passed down from `hv-root` to `hv-screen`.

| `<list>` | `<section-list>` |
|-------|-----------|
|
![list](https://github.com/Instawork/hyperview/assets/309515/6f81f420-0195-49de-b183-b36ad5b89127)
|
![sectionlist](https://github.com/Instawork/hyperview/assets/309515/0f184448-3af3-4e6e-ac3f-1c4db6fd0766)
|

Captures above tested in the demo app with the following local changes:

<details>

```diff
diff --git a/demo/src/HyperviewScreen.js b/demo/src/HyperviewScreen.js
index e91c8f9..7669382 100644
--- a/demo/src/HyperviewScreen.js
+++ b/demo/src/HyperviewScreen.js
@@ -9,6 +9,7 @@
 import React, { PureComponent } from 'react';
 import HandleBack from './HandleBack';
 import Hyperview from 'hyperview';
+import { RefreshControl } from 'react-native';
 import moment from 'moment';
 import { MAIN_STACK_NAME, MODAL_STACK_NAME } from './constants';

@@ -63,6 +64,10 @@ export default class HyperviewScreen extends PureComponent {
     });
   }

+  Refresh = () => (
+    <RefreshControl tintColor="red" />
+  )
+
   render() {
     const entrypointUrl = this.props.route.params?.url;

@@ -79,6 +84,7 @@ export default class HyperviewScreen extends PureComponent {
           openModal={this.openModal}
           push={this.push}
           route={this.props.route}
+          refreshControl={this.Refresh}
         />
       </HandleBack>
     );
```
</details>

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>