feature(deep link) - provide deep link support for navigation (#691)
This is a working prototype of deep links with custom navigators.

- For stack navigators, the custom navigator provides an injection of a
custom router which overrides the behavior of initial state and updates
for new routes
  - In the override, all routes are pushed onto the stack
- For tab navigators, the custom navigator calls a 'navigate' whenever
the initial route name prop changes

This functionality requires mobile branch
https://github.com/Instawork/mobile/tree/hardin/deep-link-url-wip

**Tabs**
With the default behavior, the initial state is correct: shifts-route
(middle tab), and second-shifts (right sub tab). However, the deep link
has no effect. The current tab selection remains.

Deep link
```xml
<doc {% namespaces %}>
  <navigator id="root-navigator" type="stack" merge="true">
    <nav-route id="tabs-route">
      <navigator id="tabs-navigator" type="tab" merge="true">
        <nav-route id="shifts-route" href="{% url 'biz-app-sub-navigator' %}" selected="true">
          <navigator id="shifts" type="tab" merge="true">
            <nav-route id="second-shifts" href="{% url 'biz-app-shift-group-list' %}" selected="true"/>
          </navigator>
        </nav-route>
      </navigator>
    </nav-route>
  </navigator>
</doc>
```

| initial state | deep link |
| -- | -- |
|
![initstate](https://github.com/Instawork/hyperview/assets/127122858/9fcb9021-eb6c-40cb-a3f9-ba7759a678d5)
|
![deeplink](https://github.com/Instawork/hyperview/assets/127122858/3bf36ccd-50aa-4bb7-8c9b-2167d8ac2512)
|

With an override navigator, the deep link is able to change the
selection in both the main tab and the sub-tab navigators.
| initial state | deep link |
| -- | -- |
|
![initstate](https://github.com/Instawork/hyperview/assets/127122858/9fcb9021-eb6c-40cb-a3f9-ba7759a678d5)
|
![deeplink](https://github.com/Instawork/hyperview/assets/127122858/9fd88766-cb00-4880-9ff0-849c3cc9ded1)
|

**Stacks**
With the default behavior, the stack navigator pushes only one screen
onto the stack: the one with a naming matching the "initialRouteName"
property value. Deep links provide no additional functionality.

Initial state (index.xml)
```xml
<doc {% namespaces %}>
  <navigator id="root-navigator" type="stack">
    <nav-route id="tabs-route">
      <navigator id="tabs-navigator" type="tab">
        <nav-route id="live-shifts-route" href="{% url 'biz-app-hub' %}"/>
        <nav-route id="shifts-route" href="{% url 'biz-app-sub-navigator' %}" selected="true"/>
        <nav-route id="account-route" href="{% url 'biz-app-account' %}" selected="true"/>
      </navigator>
    </nav-route>
    <nav-route id="company" href="{% url 'biz-app-account-company' %}" modal="true"/>
  </navigator>
</doc>
```

Deep link
```xml
<doc {% namespaces %}>
  <navigator id="root-navigator" type="stack" merge="true">
    <nav-route id="account" href="{% url 'biz-app-account' %}" modal="true"/>
  </navigator>
</doc>
```

Given this initial state, the desired outcome is to have the "company"
modal on top of the tabs.

| initial state | deep link |
| -- | -- |
|
![initstate](https://github.com/Instawork/hyperview/assets/127122858/22b801b8-b3c6-45d4-a91a-ce93dea950c0)
|
![deeplink](https://github.com/Instawork/hyperview/assets/127122858/b30fe8d7-c437-4970-a4bc-9d146107d84b)
|

After overriding the navigator and router, the stack is able to push all
screens onto the stack in the initial state and is able to push screens
onto the stack when a deep link is received.

NOTE: There is currently a bug causing the stack initial state to open
and immediately close the modal screen. It is caused by the overrides of
tabs and stack working against each other. A separate task will be
created to look into fixes.

| initial state | deep link |
| -- | -- |
|
![initstate](https://github.com/Instawork/hyperview/assets/127122858/b95b81f4-68ec-4bad-a728-0c7a15604fb9)
|
![deeplink](https://github.com/Instawork/hyperview/assets/127122858/370e9d77-a8b2-4216-b5a6-63c05bd08e0e)
|


Asana: https://app.asana.com/0/1204008699308084/1205129681586820/f