fix: deep link modals (#760)
Deep links which trigger modals were not working. The core reason is the
navigation code creates a dynamic stack navigator for every modal. This
change was not reflected in the DOM so as the deep link handling cleaned
out undefined screens, it could not match the dynamic navigator with one
in the DOM.

Changes: update the DOM to reflect the added navigator and child route.

Asana: https://app.asana.com/0/1204008699308084/1206050332658089/f