feature: add ability to block back navigation from hardware or software event (#771)
Allow developers to block the user going back or closing a screen.
Example use cases:
- User has unsaved changes
- User may have not scrolled to see all content
- User may miss important information on the screen

The implementation adds a new "back" trigger into Hyperview.

See the commits for a breakdown of the changes.

| Software back | iOS swipe | Android hardware |
| -- | -- | -- |
|
![software](https://github.com/Instawork/hyperview/assets/127122858/983faa91-e9a7-4651-ad06-ca93927b037c)
|
![iOS](https://github.com/Instawork/hyperview/assets/127122858/533af5e5-8dcd-4774-9fe1-76dd69afcd0b)
|
![android](https://github.com/Instawork/hyperview/assets/127122858/ee72b3c6-09b8-4416-82b5-0bba1a653a14)
|


**Example implementation**
In the following example, when the user attempts to go back, the top
behavior with `trigger="back"` will be called. Since there is a behavior
to handle `back`, the normal navigation is canceled. The resulting
`alert` performs two behaviors on click of the OK button: it hides the
containing view (thus removing behaviors) and then it calls an
`action="back"` which will continue the back navigation.



```XML
<view id="alert-container">
  <behavior
    trigger="back"
    action="alert"
    alert:title="You have unsaved changes. Go back?"
  >
  <alert:option alert:label="Cancel" />
  <alert:option
    alert:label="OK"
  >
    <behavior
      action="hide"
      target="alert-container"
    />
    <behavior
      action="back"
    />
  </alert:option>
  </behavior>
</view>
```

Asana: https://app.asana.com/0/1204008699308084/1205954494594099/f