Refactor ComponentRegistry into self-contained class (#882)
Previously, `ComponentRegistry` was a type that mapped namespace/local
names to component classes. This object was passed around and used for
rendering to map an XML element to a React component. It was becoming a
bit cumbersome when we introduced a second form component registry used
to collect form data. The functionality was scattered across helpers in
different files, and we had to maintain these two registries. I was also
running into issues when I wanted to expose the form registry in the
public Hyperview API so custom components can access form data.

This refactor addresses those issues by creating a new
`ComponentRegistry` class. This class handles lookups of components,
with support for aliases. It also encapsulates the logic of pulling out
form data. With this new class, we only need one registry object that
can be passed around and handle component lookups and generating form
data.

To-Do
- [x] Do we need to be backwards compatible? The component registry was
exposed to custom components as `props.options.componentRegistry`, but
it was never publicly documented. (Fixed in
https://github.com/Instawork/mobile/pull/6186)

---------

Co-authored-by: Florent Bonomo <flochtililoch+github@gmail.com>