feat(hv-navigator): Option for custom tab bar component (#950)
Add new prop `navigationComponents` to Hyperview root component, that
for now allows passing a key `TabBar` set to a React component that will
be used to render the tab bar of a bottom tab navigator. In the future,
we'll add more keys to configure other elements of the Navigation UI
such as the header of stack components, or top tab bar when we support
equivalent navigator.