feat(demo): Custom bottom tab bar component (#951)
Add a new component to render the bottom tab bar contextually with the
tab navigator. The advantages over the previous approach of rendering
the tab bar as part of the screen:
- there's a single instance of a tab bar that is used by the navigator,
instead of one instance per tab
- the tab bar is always visible even while switching to a tab that has
not loaded yet, hence removes the need for the work around to include an
additional instance in a "loading" screen.

The tab bar UI is still provided by the screen, via a new custom
element. This component does not render anything but instead lifts the
rendering props into a shared context. This context is then used by the
main custom component attached to the tab navigator to render the UI.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced a new bottom tab navigation system to enhance user
navigation.
	- Added a `BottomTabBar` component for rendering tab navigation.
- Implemented context management for dynamic tab properties with
`BottomTabBarContext`.

- **Bug Fixes**
	- Updated identifiers for tabs to improve clarity and functionality.

- **Documentation**
	- Enhanced schema definitions to support new navigation features.

- **Chores**
- Cleaned up and organized style definitions for better maintainability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>