fix(demo): Various fixes related to schema validation (#963)
Fix various issues related to schema validation, in particular due to
our set up with 11ty:
- Front matter ending up in generated XML files - this happens with
files that were both destined to be served from their own endpoints but
also included from other templates. To solve this, I created
intermediate files for those that need to be served, and remove the
front matter for those that are just included.
- I've also updated the list of files in .eleventyignore to be explicit
- this list represents all the templates that are meant to be included
but should not be served as endpoints (hence not generated as static XML
files). In the future, we will segregate private and public files to
avoid the need to define this list.

Also in this PR:
- remove exclusions from the XML validation command - now most generated
templates are validated (the templates related to the share
functionality, which uses an instawork namespace, is still ignored for
now, it will be ported to a custom behavior in a future PR)
- change the custom bottom bar component namepace prefix from
instawork.com to hyperview.org
- fix a few templates that would produce invalid XML
- fix a few schema definitions that were incorrect

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>