chore(demo): update dependencies and fix new errors (#1021)
Migrate update instructions into a script, and run it.

Testing still in progress, there are new regressions around navigation,
date picker and more that will need to be investigated.

https://app.asana.com/0/1204008699308084/1209103436957309/f

Note:
Since Hyperview core also needs to be updated to have matching
dependencies, I'm using `yarn link` locally and on CI to link the main
directory into the node_modules of the demo app. However this solution
does not work to build/run the app, so I'll revert the last commit once
we've released a new version of Hyperview, prior to merging this PR.

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>