fix(behaviors): resolve issue with broken targets (#1068)
An issue exists where a delayed update behavior can become orphaned
before it performs its action. This can present itself in two different
ways:
- a white screen or a "no content" error
- a behavior which appears to perform no action

The issue is caused when update behaviors don't specifically target a
view. Any previous behaviors which mutate the dom can orphan the
behavior making it unable to find the target view in the current dom.

To resolve the issue, every document is pre-processed when it is loaded.
Behaviors which implement "update" actions (`replace`, `replace-inner`,
`append`, `prepend`) are given a unique id attribute. When the behavior
is executed, Hyperview will find the behavior by id in the current dom
and then determine the appropriate target.

>Note: testing with mobile application data required 0-1ms to perform
the pre-processing.

The following shows the before and after of the two bugs:
1 - Shows how a render error can be triggered
```xml
<view id="broken-replace">
  <behavior
    action="replace"
    trigger="load"
    href="/hyperview/public/ui/ui-elements/text/bug/replacement.xml"
    delay="1000"
  />
  <behavior
    action="show"
    trigger="load"
    target="broken-replace"
  />
</view>
```
The `replace` action is performed first but has a delay. It caches its
reference to `<view id="broken-replace">`. The`show` behavior mutates
the dom. After the delay, the `replace` action is performed but the
cached view is no longer part of the dom and causes a render error.

2- Shows how the delay can be broken
```xml
<view id="broken-replace">
  <behavior
    action="show"
    trigger="load"
    target="broken-replace"
  />
  <behavior
    action="replace"
    trigger="load"
    href="/hyperview/public/ui/ui-elements/text/bug/replacement.xml"
    delay="1000"
  />
</view>
```
The `show` behavior runs first and mutates the dom. The `replace`
behavior is started and caches its parent `<view>` which is already
orphaned. It assigns a `_hv-timeout-id` to this view. After the delay,
when it tries to find the view by `_hv-timeout-id`, it can't find it in
the current dom.


| | Before | After |
| - | -- | -- |
| 1 |
![before-1](https://github.com/user-attachments/assets/5afc0f77-8ab9-421f-8f00-54638c7e9cee)
|
![after-1](https://github.com/user-attachments/assets/66ef015f-a13d-47f5-ade2-fe8a20fad035)
|
| 2 |
![before-2](https://github.com/user-attachments/assets/277b5c5f-a0d2-478e-bbdb-e930683d29a1)
|
![after-2](https://github.com/user-attachments/assets/97986d6f-d033-4186-8b6c-18882b8657e3)
|

[Asana](https://app.asana.com/0/1204008699308084/1209252357484428)

QA Plan
- check performance impact of changes
- full regression and automation tests

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209398827793436