fix(navigation): resolve performance issue (#1092)
The `blur`, `focus`, and `state` listeners all can perform heavy dom
mutations. Performing them at the same time the screen is rendering and
transitioning can cause delays or visual glitches. To resolve, these
listeners are delayed a small amount.

The use of `setTimeout` enables waiting for the screen to render before
mutating the dom. The `navStateMutationsDelay` experiment value is used
to enable the timeout. If the experiment is not set or has a value of
"0", no timeout is used. See
https://github.com/Instawork/hyperview/pull/1092/commits/35499d94ff3fd2b64f4dfc0c8d90deddc89b3812

| Before | After |
| -- | -- |
|
![before](https://github.com/user-attachments/assets/4563f32e-7c17-47ce-b1e9-d0eb4b50ad64)
|
![after](https://github.com/user-attachments/assets/0f1ab29e-8aed-4bf9-8ec0-c3845acd0960)
|


[Asana](https://app.asana.com/0/1204008699308084/1209403666139702)