feat: font scale props (#1129)
Refactor date-field / picker-field element to delegate to underlying
components that render `Text` components the responsibility to extract
their props and styles from the Hyperview props (element / options /
stylesheets).
This is a pre-requisite to make it easier for this other change (#37760)
to apply font scale props to Text components.

No demo yet, as the feature is broken in Expo SDK 52. See
https://github.com/expo/expo/issues/32900 and
https://github.com/facebook/react-native/issues/47499.

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>