chore(hv-doc): lift error component into hv-doc (#1145)
Lifting the error component into `hv-doc`.

Also resolved an issue where the `retry` option was faulty. The retry
was previously based on the url in state which was only set on success.
This meant that a first request failure would have no url to try. It
also meant that subsequent requests would end up retrying the previous
url instead of the current.

[Asana](https://app.asana.com/0/1204008699308084/1209771721803061/f)