chore(hv-doc): improve state stability and data (#1146)
1. Having the `onUpdate` inside the `useMemo` was causing occasional
infinite loops with external components which have `useEffect` hooks
that depend on `onUpdate` (see
[example](https://github.com/Instawork/mobile/blob/81b00f1974f373150580b9c76e113627eca312b4/core-mobile/src/components/hyperview-s3-file-upload/hooks/ui-state.ts#L38-L83)).
By moving the `onUpdate` outside of the memo, it becomes a stable
reference and does not trigger external hooks.

2. Also stabilized the `setScreenState` callback by depending on the
presence of an element rather than the element itself.

3. Temporarily provide immediate access to the `url` value of state.
Currently external components are still calling loadUrl which is
occurring before the state has completed. Once the load hooks are
internalized, this change can be removed.