chore(modal): update `isModal` param usage (#1147)
Using a navigation structure with a modal:
```
<navigator id="home" type="stack">
  <nav-route id="test" href="foo" modal="true" />
</navigator>
```
The components will automatically inject a stack navigator into the
modal, resulting in the following:
```
<navigator id="home" type="stack">
  <nav-route id="test" href="foo" modal="true">    <--- modal=true
    <navigator id="modal-stack" type="stack">
      <nav-route id="modal-test" href="foo" modal="false" />   <--- modal=false
    </navigator>
  </nav-route>
</navigator>
```
Note in this case the inner route "modal-test" is the one which performs
the load. In current Hyperview, the modal is set to false for this
route.

The change will preserve all existing functionality but ensure the
"isModal" param for both routes will be "true". The new `needsSubStack`
property will be used internally to handle the required setup, and the
`isModal` will be purely informational indicating both the original
route and the created sub-stack route both have a value of `true`.

| Before | After |
| -- | -- |
|
![before](https://github.com/user-attachments/assets/dc264dc0-8512-4253-a425-fc7b1e45cd1d)
|
![after](https://github.com/user-attachments/assets/f070238a-5d18-45de-b298-ada7eefac745)
|

[Asana](https://app.asana.com/0/1204008699308084/1209939499908844/f)