chore(hv-doc): consolidate hv-doc state (#1160)
Merging the hv-doc implementation to use a single instance across both
`hv-route` and `hv-screen`. This has the benefit of reducing individual
state management, load tracking, and removes the need to pass a clone of
the loaded doc into `hv-screen`. This also includes setting the initial
state to use `null` values instead of `undefined` to reduce the effect
of state updates causing re-renders.

This has been tested within HV Demo including features which force a
screen to reload.


[Asana](https://app.asana.com/1/47184964732898/project/1205761271270033/task/1209875001265872?focus=true)