chore(section-list): revert FC conversion (#1221)
While the section-list as Functional Component works well in the demo
and in many places in the mobile apps, it is failing in the profile
section. After spending some time trying to debug the problem, it seemed
better to postpone this FC migration for now. The other optimizations
including the new context hierarchy remain.