fix: date picker style (#1223)
New version of datepicker no longer auto center - this fix forces it
back in the center of the picker modal. This PR also cleans up
duplication on the demo styles for the picker modal.

| Before | After |
|--|--|
| ![Simulator Screenshot - iPhone 16 Pro - 2025-02-14 at 16 52
40](https://github.com/user-attachments/assets/31ee777f-47ef-47c8-92fa-db21e23d7745)
| ![Simulator Screenshot - iPhone 16 Pro - 2025-02-14 at 16 41
59](https://github.com/user-attachments/assets/d1e0b053-559a-4a00-bc9c-7ac494ab1977)
|

Note: this is a redo of #1082, which was somehow canceled out by a
force-push on the latest Expo SDK update integration branch

---------

Co-authored-by: flochtililoch <flochtililoch@gmail.com>