chore: remove support for maps on web (#1226)
The underlying 3rd party replacement for react-native-web is not
compatible with React 19, and Expo currently does not provide an out of
the box solution for maps. This functionality was added as part of
Instawork hackathon, but the feature never ended up being used, so we
wont be investing in supporting this for the time being.

Co-authored-by: flochtililoch <flochtililoch@gmail.com>