chore(hv-doc): use full url comparison (#1237)
HvDoc receives alternate versions of a url as either a full url or a
partial which can cause unnecessary reloads. This change ensures the
full url is compared in all cases to avoid reloads.

> A [separate
task](https://app.asana.com/1/47184964732898/project/1205761271270033/task/1210316546014483?focus=true)
exists to track the root cause of why the url is changing.


[Asana](https://app.asana.com/1/47184964732898/project/1204008699308084/task/1210778282863683?focus=true)