feat(sync): Add synchronization support for requests triggered in behaviors (#1243)
Adding request synchronization as [described
here](https://github.com/Instawork/hyperview/pull/736#issuecomment-3176454724).
This will help prevent issues where responses can complete out-of-order
from the requests.

Synchronization works on behaviors that make HTTP requests. Support is
added through two behaviors:
- `sync-id`: An ID used to synchronize requests. All requests with this
ID will be subject to the sync method.
- `sync-method`. One of `drop` (default) or `replace`:
- if `drop`, subsequent requests with the same sync ID will not be made
if there is a request currently in-flight
- if `replace`, subsequent requests will be made. However, only the
response of the last will be processed, the other responses will be
dropped and treated as a no-op.

### Implementation
The fetch API in React Native does not support cancellation with a
signal. So instead, we need to track the current request for a given
sync ID, and ignore responses from other requests (depending on sync
method).

- Add support for extracting and passing along the `sync-id` and
`sync-method` attributes.
- In `parser.ts`, update the logic that makes requests to support
optional synchronization.
- If `syncId` is provided, store request promises in a map, keyed by
`syncId`
- Logic based on `syncMethod` is implemented around the fetch logic.
- `drop` method is processed before the request is made, potentially
returning a `NO_OP` if there is another request in flight for `syncId`.
- `replace` method is processed when the response is received,
potentially returning a `NO_OP` if there's a newer request in flight for
`syncId`.


### Other changes
- [x] Updated demo with new "sync" screen that shows the `replace`
method in action. Uses a dynamic response that delays responses by a
random amount, to stress test the sync capabilities.
- [x] Updated schema with the new attributes.
- [ ] Update website docs