chore(hyper-ref): remove legacy bridge (#1253)
The path `src/core/hyper-ref` was a legacy path which existed before the
code reorganization. This file was create as a temporary bridge to allow
the demo and mobile clients to work without migrating. The path has been
corrected in both demo and mobile apps and this legacy path can be
removed.


[Asana](https://app.asana.com/1/47184964732898/project/1205761271270033/task/1210633057805216?focus=true)