chore(error): improve error logging context (#1262)
Improving `ParserError` and `XMLRequiredElementNotFound` errors in
Sentry by injecting additional values including url, status, and
content.

In the examples below, look for the "Additional Data" section near the
bottom. Also note that the `content` may be filtered by Sentry if it
contains restricted data.

The content being passed is trimmed and centered on where the error
occurred. For parsing errors, this will be at the location where the
parsing failed (line, col). For xml elements, this will be at the
closing tag of the expected parent.

| Parsing | Element |
| -- | -- |
|
![parsing](https://github.com/user-attachments/assets/b346a0b3-53d9-466f-804c-a37899ef8fb4)
|
![element](https://github.com/user-attachments/assets/b903107f-4b21-4afb-b9db-84afe11f70c3)
|

## Parser error
- before:
https://instawork.sentry.io/issues/6929029873/?project=1549809&query=is%3Aunresolved%20parsererror&referrer=issue-stream
- after:
https://instawork.sentry.io/issues/6947071566/?query=-temp&referrer=issue-stream

## XMLRequiredElementNotFound
- before:
https://instawork.sentry.io/issues/6828134611/?project=1549809&query=is%3Aunresolved%20xmlrequired&referrer=issue-stream
- after:
https://instawork.sentry.io/issues/6947069133/?query=-temp&referrer=issue-stream

The contents of the on-screen debug error have also been improved for
the parser error to show the relevant snippet.

| Before | After |
| -- | -- |
|
![before-parser](https://github.com/user-attachments/assets/a46034a7-fbe8-4675-b7d1-0a9ceed6a327)
|
![after-parser](https://github.com/user-attachments/assets/dbe7535e-6df3-4e5a-a8fd-fc46a4e3a9ab)
|



[Asana](https://app.asana.com/1/47184964732898/project/1204008699308084/task/1211614396813869?focus=true)