chore(render): degrade log level (#1275)
Hyperview is currently logging warnings when trying to render components
which are not present. Given that it will gracefully handle the
situation by not trying to render the component, this is being
downgraded to an `info` event.

[Related Sentry
events](https://instawork.sentry.io/issues/6805106942/?project=1549809&query=is%3Aunresolved%20attribute&referrer=issue-stream&sort=freq)


[Asana](https://app.asana.com/1/47184964732898/project/1204008699308084/task/1212311671958494?focus=true)